app/src/test/java/com/example/moviecatalogue/presentation/detail/DetailViewModelTest.kt
package com.example.moviecatalogue.presentation.detail

import androidx.arch.core.executor.testing.InstantTaskExecutorRule
import androidx.lifecycle.MutableLiveData
import androidx.lifecycle.Observer
import com.example.moviecatalogue.data.MovieRepository
import com.example.moviecatalogue.data.source.local.entity.Movie
import com.example.moviecatalogue.data.source.local.entity.TvShow
import com.example.moviecatalogue.utils.FakeData
import com.nhaarman.mockitokotlin2.verify
import org.junit.Before
import org.junit.Test

import org.junit.Assert.*
import org.junit.Rule
import org.junit.runner.RunWith
import org.mockito.Mock
import org.mockito.Mockito.`when`
import org.mockito.junit.MockitoJUnitRunner

@RunWith(MockitoJUnitRunner::class)
class DetailViewModelTest {

    private lateinit var viewModel: DetailViewModel
    private val dummyMovie = FakeData.getDummyRemoteMovies()[0]
    private val movieId = dummyMovie.id
    private val dummyTvShow = FakeData.getDummyRemoteTvShows()[0]
    private val tvShowId = dummyTvShow.id

    @get:Rule
    var instantTaskExecutorRule = InstantTaskExecutorRule()

    @Mock
    private lateinit var movieRepository: MovieRepository

    @Mock
    private lateinit var movieObserver: Observer<Movie>

    @Mock
    private lateinit var tvShowObserver: Observer<TvShow>

    @Before
    fun setUp() {
        viewModel = DetailViewModel(movieRepository)

        viewModel.setSelectedMovie(movieId!!)
        viewModel.setSelectedTvShow(tvShowId!!)
    }

    @Test
    fun getMovieDetail_completed_returnTheRightData() {
        val movie = MutableLiveData<Movie>()
        movie.value = dummyMovie

        `when`(movieRepository.getMovieDetail(movieId!!)).thenReturn(movie)
        val movieEntity = viewModel.getMovie().value as Movie
        verify(movieRepository).getMovieDetail(movieId)
        assertNotNull(movieEntity)
        assertEquals(dummyMovie.id, movieEntity.id)
        assertEquals(dummyMovie.backdropPath, movieEntity.backdropPath)
        assertEquals(dummyMovie.title, movieEntity.title)
        assertEquals(dummyMovie.overview, movieEntity.overview)

        viewModel.getMovie().observeForever(movieObserver)
        verify(movieObserver).onChanged(dummyMovie)
    }

    @Test
    fun getTvShowDetail_completed_returnTheRightData() {
        val tvShow = MutableLiveData<TvShow>()
        tvShow.value = dummyTvShow

        `when`(movieRepository.getTvShowDetail(tvShowId!!)).thenReturn(tvShow)
        val tvShowEntity = viewModel.getTvShow().value as TvShow
        verify(movieRepository).getTvShowDetail(tvShowId)
        assertNotNull(tvShowEntity)
        assertEquals(dummyTvShow.id, tvShowEntity.id)
        assertEquals(dummyTvShow.backdropPath, tvShowEntity.backdropPath)
        assertEquals(dummyTvShow.name, tvShowEntity.name)
        assertEquals(dummyTvShow.overview, tvShowEntity.overview)

        viewModel.getTvShow().observeForever(tvShowObserver)
        verify(tvShowObserver).onChanged(dummyTvShow)
    }
}
